+++ includes/shortcodes/year.php
@@ -2,12 +2,12 @@
 /**
  * Creates [year] shortcode for the theme
  *
- * @package CassidyDC\BlockTheme\Functions
+ * @package CassidyDC\StarterBlockTheme\Functions
  * @version 1.0.0
  */
 
 declare( strict_types = 1 );
-namespace CassidyDC\BlockTheme;
+namespace CassidyDC\StarterBlockTheme;
 
 /**
  * Displays the current year