+++ patterns/boxes/box-default.php
@@ -1,12 +1,12 @@
 <?php
 /**
  * Title: Box - Default
- * Slug: cassidydc-block-theme/box-default
+ * Slug: cassidydc-starter-block-theme/box-default
  * Categories: boxes
  * Keywords: boxes, default box, box
  * Viewport Width: 1440
  *
- * @package CassidyDC\BlockTheme\Patterns
+ * @package CassidyDC\StarterBlockTheme\Patterns
  * @since 1.0.0
  * @version 1.0.0
  */